Separated pennychat model to new app
