+++ .gitignore
@@ -8,3 +8,4 @@ venv
 .idea
 .DS_Store
 .vscode
+staticfiles/
