+++ penny_university/settings.py
@@ -38,7 +38,8 @@ INSTALLED_APPS = [
     'django.contrib.messages',
     'django.contrib.staticfiles',
     'bot.apps.BotConfig',
-    'home'
+    'home',
+    'pennychat'
 ]
 
 MIDDLEWARE = [
